pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "virtual-ops-manager"
version = "1.0.0"
description = "OpenEnv environment simulating an enterprise Virtual Operations Manager"
authors = [
    {name = "Jnapika Pilli", email = "pillijnapika8@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
keywords = ["openenv", "rl", "agent", "customer-support", "email-triage"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.7.0",
    "openai>=2.7.2",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[project.urls]
Homepage = "https://huggingface.co/spaces/jnapika/virtual-ops-manager"
Repository = "https://huggingface.co/spaces/jnapika/virtual-ops-manager"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["frontend*", "tests*", "*.egg-info"]

[tool.openenv]
name = "virtual-ops-manager"
version = "1.0.0"
entry_point = "server.app:app"
reset_endpoint = "/api/reset"
step_endpoint = "/api/step"
state_endpoint = "/api/state"
status_endpoint = "/api/status"
tasks = ["easy_01", "medium_01", "hard_01", "expert_01", "expert_02"]